Installing sorted by "dependency required"
Instead of building the dependency graph, we count how many dependencies
a package has, and install the ones with fewer before than the others.
This list is always updated. CPU-expensive, but simple.
This works well if there are no cycles. (one at a time).
If we have cycle, it should install these the-right-way